Skip sync if the dd_tag already exists
This is a safety mechanism to ensure we don't override an existing weekly tag if the head of the datadog branch changes.
The conditionals used on every subsequent steps after the early exit are needed because github actions do not have proper support for early exit with success. See https://github.com/orgs/community/discussions/82744